Remove require call in in favour of import statment
In StructuredText/index.ts vue was required via `require('vue')` during
the execution of `appendKeyToValidElement`. When this component is build
via Vite, the require call is not resolved and throws an error in the
produciton build.

The `vue-demi` package (which is `import`ed already) does export the
`isVNode` and `cloneVNode` functions (which is the reason vue was
required). The `vue-demi` versions are used now